Zadanie13/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
cloudinary
pydantic
python-dotenv
fastapi-limiter
email-validator
python-jose
